implement (<**>) using (=<<) and (<$>)
This version better reflects the documentation on L78:
```-- | Witness that all things with (=<<) and (<$>) also have (<*>).```

I was stumped by this exercise while going through https://github.com/system-f/fp-course
and came here looking for a solution. Eventually I figured out this formulation, which I think
is more fitting. It's actually not immediately obvious to me (a beginner) that `(>>=)` is
defined here.